Separate list divisors
